Add PIR motion sensor Arduino sketch
Created pir_motion_sensor.ino to detect human motion using a PIR sensor. 
When motion is detected, the onboard LED connected to pin 13 lights up and a message is printed to the Serial Monitor. 
The sensor output is read on digital pin 2. Includes setup and loop functions with serial debugging.